Create A project by Fritz
Grading students is a fundamental aspect in the workload of professors and 
teachers. It is a straightforward calculation that only involves the student’s grades per 
assessment and the subject’s grading system. In this regard, such job is relatively easy 
and quick. However, since professors and teachers deal with tens, and sometimes even 
hundreds of students, grading them becomes much more tiring and time-consuming. 
Because of this, it is essential for the teaching profession to have a trustworthy and quick 
grading program to grade students conveniently and easily. Fortunately, C programming 
language has the ability to resolve this problem. Through the use of matrix multiplication, 
a program may be developed that can easily and quickly solve student’s final grades.
The program particularly responds to the stated problem and even 
added another feature. This program has two distinct features: solving students’
final grades and comparing changes in students’ final grades as a result of grading system 
adjustments. If the professor wishes to only solve the final grades, the grading system 
variation can be set to 1. On the other hand, if he/she wishes to determine how the final 
grades of students changes when a variation in the grading system is applied, he/she 
may input the desired number of variations.